requirements.txt
absl_py==2.1.0
datasets==3.0.2
flax==0.10.2
immutabledict==4.2.1
jax==0.5.3
jaxtyping==0.3.2
matplotlib==3.10.3
mock==5.2.0
nltk==3.9.1
numpy==2.2.5
optax==0.2.4
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.3
spacy==3.8.2
tokenizers==0.21.1
torch==2.5.1+cu118
tqdm==4.67.1
transformers==4.43.3
